src/scripts/download_kaggle_data/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-storage
kaggle
flask
pandas
